\name{installGui}
\Rdversion{1.1}
\alias{installGui}
\title{
Function to start up the installer for the standalone Gui.
}
\description{
Starts the installer for the standalone version of RSiena. Only for Windows.
}
\usage{
installGui()
}
\value{ None.
}
\references{See \url{http://www.stats.ox.ac.uk/~snijders/siena/}}
\author{Ruth Ripley }

\examples{
\dontrun{installGui()}
}
\keyword{ misc }
