\name{s501}
\alias{s501}
\docType{data}
\title{Network 1 data: excerpt from 'Teenage Friends and Lifestyle Study' data.}
\description{
First timepoint network data from an excerpt of 50 girls from the
Teenage Friends and Lifestyle Study data set. Useful as a small example
of network and behaviour, for which models can be fitted quickly.
}
\format{
 The adjacency matrix for the network at time point 1.
}
\source{
  s50-network1.dat from Stocnet
}
\references{See \url{http://www.stats.ox.ac.uk/~snijders/siena/}}
\keyword{datasets}
