\name{s50a}
\alias{s50a}
\docType{data}
\title{Alcohol use data: excerpt from 'Teenage Friends and Lifestyle Study' data}
\description{
Data from an excerpt of 50 girls from the Teenage Friends and Lifestyle Study data
set. Useful as a small example of network and behaviour, for which
models can be fitted quickly.
}
\format{
 A matrix of variables relating to the use of alcohol for the actors in
 the network. Two columns, one for each time point. Coding is 1--5, high
 values indicating higher consumption.
}
\source{part of s50-alcohol.dat from the Stocnet package.
}
\references{See \url{http://www.stats.ox.ac.uk/~snijders/siena/}}
\keyword{datasets}
