\name{tmp3}
\alias{tmp3}
\docType{data}
\title{van de Bunt's Freshman dataset, time point 3}
\description{
Third timepoint of van de Bunt's freshman dataset.
}
\format{
 Adjacency matrix for the network at time point 3.
}
\source{
  vrnd32t3.dat from Stocnet
}
\references{See \url{http://www.stats.ox.ac.uk/~snijders/siena/}}
\keyword{datasets}
