\name{tmp4}
\alias{tmp4}
\docType{data}
\title{van de Bunt's Freshman dataset, time point 4}
\description{
Fourth timepoint of van de Bunt's freshman dataset.\cr

Codes: 1 = best friendship; 2 = friendship; 3 = friendly relationship;
4 = neutral relationship; 5 = troubled relationship; 0 = unknown person.
}
\format{
  Adjacency matrix for the network at time point 4.
}
\source{
  vrnd32t4.dat from
  \url{http://www.stats.ox.ac.uk/~snijders/siena/vdBunt_data.zip}
}
\references{
  Van de Bunt, G.G., M.A.J. van Duijn, and T.A.B. Snijders (1999).
  Friendship networks through time: An actor-oriented statistical network model.
  \emph{Computational and Mathematical Organization Theory}, 5, 167-192. \cr
  Also see \url{http://www.stats.ox.ac.uk/~snijders/siena/vdBunt_data.htm}.
}
\seealso{\code{\link{tmp3}}}
\keyword{datasets}
