\name{installGui}
\Rdversion{1.1}
\alias{installGui}
\title{
Obsolete function to start up the installer for the standalone Gui.
}
\description{
Once started the installer for the standalone version of RSiena. Only
for Windows. Not possible since R version 2.12.0. Use \code{siena01Gui}
within R instead.
}
\usage{
installGui()
}
\value{ None.
}
\references{See \url{http://www.stats.ox.ac.uk/~snijders/siena/}}
\author{Ruth Ripley }

\examples{
\dontrun{installGui()}
}
\keyword{ misc }
