\name{s502}
\alias{s502}
\docType{data}
\title{Network 2 data: excerpt from 'Teenage Friends and Lifestyle Study' data.}
\description{
Second timepoint network data from an excerpt of 50 girls from the
Teenage Friends and Lifestyle Study data set. Useful as a small example
of network and behaviour, for which models can be fitted quickly.
}
\format{
 Adjacency matrix for the network at time point 2.
}
\source{
 s50-network2.dat from Stocnet
}
\references{See \url{http://www.stats.ox.ac.uk/~snijders/siena/}}
\keyword{datasets}
