\name{s502}
\alias{s502}
\docType{data}
\title{Network 2 data: excerpt from "Teenage Friends and Lifestyle Study" data.
}
\description{
Second timepoint network data from   
an excerpt of the network, alcohol consumption, and smoking data     
for 50 randomly chosen girls from the
Teenage Friends and Lifestyle Study data set. Useful as a small    
example of network and behaviour, for which models can be    
fitted quickly, and for which there are no missing values.
}
\format{
 The adjacency matrix for the network at time point 2.
}
\source{
\url{https://www.stats.ox.ac.uk/~snijders/siena/s50_data.zip}
}
\references{
West, P. and Sweeting, H. (1995), \emph{Background Rationale and Design of the
West of Scotland 11-16 Study}.
Working Paper No. 52. MRC Medical Sociology Unit Glasgow.

See \url{https://www.stats.ox.ac.uk/~snijders/siena/s50_data.htm}
}
\seealso{\code{\link{s501}}, \code{\link{s503}}, \code{\link{s50a}}, 
\code{\link{s50s}}, \code{\link{s50}}}
\keyword{datasets}
