\name{tmp4}
\alias{tmp4}
\docType{data}
\title{van de Bunt's Freshman dataset, time point 4}
\description{
Fourth timepoint of van de Bunt's freshman dataset
}
\format{
  Adjacency matrix for the network at time point 4.
}
\source{
  see siena web page
}
\references{See \url{http://www.stats.ox.ac.uk/~snijders/siena/}}
\keyword{datasets}
