\name{tmp4}
\alias{tmp4}
\docType{data}
\title{van de Bunt's Freshman dataset, time point 4}
\description{
Fourth timepoint of van de Bunt's freshman dataset.\cr

Codes: 1 = best friendship; 2 = friendship; 3 = friendly relationship;  
4 = neutral relationship; 5 = troubled relationship; 0 = unknown person.
}
\format{
  Adjacency matrix for the "at least friendly relationship"
  network at time point 4.
}
\source{
  vrnd32t4.dat from
  \url{https://www.stats.ox.ac.uk/~snijders/siena/vdBunt_data.zip}
}
\references{
  Van de Bunt, G.G., van Duijn, M.A.J., and Snijders, T.A.B.  (1999),
  Friendship networks through time: An actor-oriented statistical network model.
  \emph{Computational and Mathematical Organization Theory}, 5, 167--192. 
  
  Also see \url{https://www.stats.ox.ac.uk/~snijders/siena/vdBunt_data.htm}.
}
\seealso{\code{\link{tmp3}}}
\keyword{datasets}
